Just added stuff to be inserted in database. Some stuff needs to be done for this to happen.

+++ BlocParty/app/src/main/java/com/ngynstvn/android/blocparty/api/model/database/table/PostItemTable.java
@@ -71,7 +71,7 @@ public class PostItemTable extends Table {
             return this;
         }
 
-        public Builder setPostPublishDate(int value) {
+        public Builder setPostPublishDate(long value) {
             contentValues.put(COLUMN_POST_PUBLISH_DATE, value);
             return this;
         }
